requirements.txt
openai==1.58.1
python-dotenv==1.0.1
langchain==0.2.16
langchain-openai==0.1.25
langchain-community==0.2.16
langchainhub==0.1.20
pypdf==4.3.1
faiss-cpu==1.13.2
PyMuPDF==1.26.7
telebot==0.0.5